requirements_consumer.txt
kafka-python
hdfs[kerberos] # Cài đặt hdfs với hỗ trợ kerberos (kể cả khi không dùng kerberos vẫn nên có)
# Hoặc snakebite-py3 nếu bạn chọn dùng thư viện đó